Upgraded to solrj 3.6 and optimized the server connection settings

+++ core/app/core/indexing/IndexingService.scala
@@ -7,6 +7,7 @@ import play.api.Logger
 import org.apache.solr.common.SolrInputDocument
 import org.joda.time.DateTime
 import org.joda.time.format.DateTimeFormat
+import org.apache.solr.client.solrj.SolrQuery
 
 /**
  * Indexing API
@@ -88,10 +89,16 @@ object IndexingService extends SolrServer {
   def deleteOrphansBySpec(orgId: String, spec: String, startIndexing: DateTime) {
     val fmt = DateTimeFormat.forPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
     val deleteQuery = SPEC + ":" + spec + " " + ORG_ID + ":" + orgId + " timestamp:[* TO " + fmt.print(startIndexing) + "]"
-    Logger.info("Deleting orphans from dataset from Solr Index: %s".format(deleteQuery))
-    val deleteResponse = getStreamingUpdateServer.deleteByQuery(deleteQuery)
-    deleteResponse.getStatus
-    getStreamingUpdateServer.commit
+    val numFound = getSolrServer.query(new SolrQuery(deleteQuery)).getResults.getNumFound
+    if (numFound > 0) {
+      val deleteResponse = getStreamingUpdateServer.deleteByQuery(deleteQuery)
+      deleteResponse.getStatus
+      getStreamingUpdateServer.commit
+      Logger.info("Deleting orphans %s from dataset from Solr Index: %s".format(numFound.toString, deleteQuery))
+    }
+    else
+      Logger.info("No orphans found for dataset in Solr Index: %s".format(deleteQuery))
+
   }
 
 }
